Added a 'reseredBufferSpace' count in Chanend.h/cpp to provide *some* simulation of messages in flight. This avoids messages being sent when there is not sufficient buffer space, but *only* when a route is open between the sender and receiver.
When the route is closed during an exchange, futher messages can be sent before
the closing token has arrived and assume the route has opened, threreby causing
a similar problem.  The effect is that messages following an END do not cause
the receiving thread to be scheduled.

This is only a problem with message sequences that do not adhere to the XC
protocol.